# Commerce-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeRoute B2B marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Admin — list all users
      parameters:
        - in: query
          name: role
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}/role:
    post:
      operationId: updateUserRole
      tags: [users]
      summary: Admin — change a user's role
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/me/payment:
    patch:
      operationId: updatePaymentCredentials
      tags: [users]
      summary: Update own payment credentials (manufacturer / any seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    Role:
      type: string
      enum: [shopkeeper, manufacturer, transporter, admin]
    IdProofType:
      type: string
      enum: [aadhaar, pan, passport, drivers_license, voter_id]
    User:
      type: object
      required: [id, name, email, role, companyName, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        companyName:
          type: string
        phone:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        idProofType:
          $ref: "#/components/schemas/IdProofType"
          nullable: true
        idProofNumber:
          type: string
          nullable: true
        gstin:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        upiId:
          type: string
          nullable: true
        bankAccount:
          type: string
          nullable: true
        bankIfsc:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    UpdateUserRoleInput:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/Role"
...
    UpdatePaymentInput:
      type: object
      properties:
        upiId:
          type: string
          nullable: true
        bankAccount:
          type: string
          nullable: true
        bankIfsc:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
...